+++ NEWS.md
@@ -1,4 +1,5 @@
 # amapGeocode 0.2.0
+* Add other two functions: `getAdmin` and `extractAdmin`
 
 # amapGeocode 0.1.0